Tesaurus dels topònims del CTV amb elements d'OSM

Inclou els elements d'OSM que es troben al mateix municipi que el topònim del CTV i que tenen el mateix nom (name o name:ca) sense tenir en compte les majúscules i minúscules.

Ús

data("osm_CTV")

Format

Un data.frame amb 14314 files amb les següents 55 variables.

toponim

Topònim de l'AVL.

name

Etiqueta name de l'objecte a OSM.

name:ca

Etiqueta name:ca de l'objecte a OSM.

alt_name

Etiqueta alt_name de l'objecte a OSM.

alt_name:ca

Etiqueta alt_name:ca de l'objecte a OSM.

avl_tipus

Tipus d'element segons el CTV (Altres llocs d'interés, Hidrografia, Orografia, Partides i paratges, Poblament, Relleu litoral i marí, Vies de comunicació).

wikidata

Etiqueta wikidata de l'objecte a OSM.

ref:avl

Número de pàgina, columna i fila al pdf original.

osm_id

Identificador numèric de l'objecte a OSM.

osm_type

Tipus de l'objecte a OSM.

osm_municipi

Nom del municipi on es troba l'element a OSM (name:ca).

osm_comarca

Nom de la comarca on es troba l'element a OSM (name:ca).

osm_typeMun

Identificador numèric del municipi on es troba l'element a OSM.

osm_idMun

Tipus del municipi on es troba l'element a OSM.

osm_center_lon

Longitud del centre de la caixa de l'objecte a OSM.

osm_center_lat

Latitud del centre de la caixa de l'objecte a OSM.

avl_municipi

Nom dels municipis segons el Corpus Toponímic Valencià.

...

Etiquetes dels elements de OSM amb més de 100 casos entre tots els elements.

% \item{\code{amenity}}{a character vector} % \item{\code{boundary}}{a character vector} % \item{\code{bridge}}{a character vector} % \item{\code{building}}{a character vector} % \item{\samp{cycleway:both}}{a character vector} % \item{\code{ele}}{a character vector} % \item{\code{highway}}{a character vector} % \item{\code{historic}}{a character vector} % \item{\code{intermittent}}{a character vector} % \item{\code{landuse}}{a character vector} % \item{\code{lane_markings}}{a character vector} % \item{\code{lanes}}{a character vector} % \item{\code{layer}}{a character vector} % \item{\code{lit}}{a character vector} % \item{\code{man_made}}{a character vector} % \item{\code{maxspeed}}{a character vector} % \item{\samp{name:es}}{a character vector} % \item{\code{natural}}{a character vector} % \item{\code{oneway}}{a character vector} % \item{\code{place}}{a character vector} % \item{\code{population}}{a character vector} % \item{\code{public_transport}}{a character vector} % \item{\code{ref}}{a character vector} % \item{\samp{ref:ine}}{a character vector} % \item{\code{religion}}{a character vector} % \item{\code{sidewalk}}{a character vector} % \item{\code{source}}{a character vector} % \item{\samp{source:ele}}{a character vector} % \item{\samp{source:name}}{a character vector} % \item{\code{surface}}{a character vector} % \item{\code{tracktype}}{a character vector} % \item{\code{tunnel}}{a character vector} % \item{\code{type}}{a character vector} % \item{\code{waterway}}{a character vector} % \item{\samp{waterway_category:NDS}}{a character vector} % \item{\samp{waterway:type}}{a character vector} % \item{\code{width}}{a character vector} % \item{\code{wikipedia}}{a character vector}

Nota

Tots els topònims són en valencià excepte els de l'ALCÚDIA DE VEO, que n'hi ha alguns en castellà.

Referències

Dades generades a data-raw/tesaurus_OSM-CTV.R.

Vegeu també

Exemples

if (FALSE) { # \dontrun{
mapa <- sf::st_as_sf(
  osm_CTV,
  coords = c("osm_center_lon", "osm_center_lat"), remove = FALSE,
  crs = "+proj=longlat +datum=WGS84 +no_defs"
)
mapa
} # }